change latest-news to ts
