--- src/features/latest-news/redux/middleware.js
@@ -1,14 +1,0 @@
-import { BOOTSTRAP } from "../../../store/constants";
-import { updateLatestNewsAction } from "./actions";
-
-export const latestNewsMiddleware = ({
-  socketService
-}) => store => next => action => {
-  if (action.type === BOOTSTRAP) {
-    const socket = socketService.get();
-    socket.on("latestNews", payload => {
-      store.dispatch(updateLatestNewsAction(payload));
-    });
-  }
-  return next(action);
-};
