+++ src/features/latest-news/redux/reducer.ts
@@ -5,7 +5,7 @@ import { Reducer } from "redux";
 import { ResetAction } from "../../../store/actions";
 
 export type LatestNewsState = {
-  selectedLatestNews: LatestNewsData | null;
+  selectedLatestNews: LatestNewsData[] | null;
 };
 
 const initialState = { selectedLatestNews: null };
